refactor: #11 block, page 단위 컴포넌트 분할
