--- src/blocks/AvatarGroup/index.tsx
@@ -1,7 +1,0 @@
-import AvatarGroup from './AvatarGroup'
-
-export default function AvatarGroupContainer() {
-    return (
-        <AvatarGroup />
-    )
-}
