--- src/blocks/CardGrid/index.tsx
@@ -1,7 +1,0 @@
-import CardGrid from "./CardGrid";
-
-export default function CardGridContainer() {
-    return (
-        <CardGrid />
-    )
-}
