--- src/blocks/SearchTextField/index.tsx
@@ -1,5 +1,0 @@
-import SearchTextField from './SearchTextField';
-
-export default function SearchTextFieldContainer() {
-  return <SearchTextField />;
-}
