--- src/blocks/SplitButton/index.tsx
@@ -1,7 +1,0 @@
-import SplitButton from "./SplitButton";
-
-export default function SplitButtonContainer() {
-    return (
-        <SplitButton />
-    )
-}
